# docs/openapi/codewithmee-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "AuditState": {
        "additionalProperties": false,
        "properties": {
          "authorityRevision": {
            "minimum": 1,
            "type": "integer"
          },
          "organizationRole": {
            "enum": [
              "admin",
              "analyst",
              "grader",
              "instructor",
              "owner"
            ],
            "type": "string"
          },
          "ownerUserId": {
            "format": "uuid",
            "type": "string"
          },
          "platformRole": {
            "enum": [
              "learner",
              "moderator",
              "superadmin",
              "support"
            ],
            "type": "string"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "status": {
            "enum": [
              "active",
              "banned",
              "deletion_pending",
              "suspended"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "AuditEvent": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "minLength": 1,
            "type": "string"
          },
          "actorSessionId": {
            "type": [
              "string",
              "null"
            ]
          },
          "actorUserId": {
            "type": [
              "string",
              "null"
            ]
          },
          "afterState": {
            "$ref": "#/components/schemas/AuditState"
          },
          "beforeState": {
            "$ref": "#/components/schemas/AuditState"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          },
          "operatorReference": {
            "type": [
              "string",
              "null"
            ]
          },
          "organizationId": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "requestId": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "enum": [
              "api",
              "bootstrap_cli"
            ],
            "type": "string"
          },
          "targetUserId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "action",
          "actorSessionId",
          "actorUserId",
          "afterState",
          "beforeState",
          "id",
          "occurredAt",
          "operatorReference",
          "organizationId",
          "reason",
          "requestId",
          "source",
          "targetUserId"
        ],
        "type": "object"
      },
      "AuthorityUser": {
        "additionalProperties": false,
        "properties": {
          "authorityRevision": {
            "minimum": 1,
            "type": "integer"
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "displayName": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "platformRole": {
            "enum": [
              "learner",
              "moderator",
              "superadmin",
              "support"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "banned",
              "deletion_pending",
              "suspended"
            ],
            "type": "string"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "authorityRevision",
          "avatarUrl",
          "createdAt",
          "displayName",
          "email",
          "emailVerified",
          "id",
          "platformRole",
          "status",
          "username"
        ],
        "type": "object"
      },
...
      "OwnershipMembership": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "organizationId": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "analyst",
              "grader",
              "instructor",
              "owner"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "revoked",
              "suspended"
            ],
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "joinedAt",
          "organizationId",
          "role",
          "status",
          "userId"
        ],
        "type": "object"
      },
      "OwnershipOrganization": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ownerUserId": {
            "format": "uuid",
            "type": "string"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "ownerUserId",
          "revision"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "type": "string"
          },
          "detail": {
            "maxLength": 1000,
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "maxItems": 20
          },
          "instance": {
            "maxLength": 500,
            "type": "string"
          },
          "meta": {
            "additionalProperties": false,
            "properties": {
              "maxBytes": {
                "minimum": 1,
                "type": "integer"
              }
            },
            "type": "object"
          },
          "requestId": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "title",
          "type"
        ],
        "type": "object"
      },
...
      "ValidationIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "additional_property",
              "constant",
              "enum",
              "format",
              "max_items",
              "max_length",
              "max_properties",
              "min_items",
              "min_length",
              "min_properties",
              "minimum",
              "maximum",
              "one_of",
              "pattern",
              "required",
              "type"
            ],
            "type": "string"
          },
          "pointer": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "code",
          "pointer"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "csrfHeader": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "refreshCookie": {
        "in": "cookie",
        "name": "cwm_refresh",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Implemented CodeWithMee v1 operations. Unknown request fields are rejected. Future roadmap operations are intentionally omitted until deployed.",
    "title": "CodeWithMee API",
    "version": "1.0.0"
  },
...
  "openapi": "3.1.1",
  "paths": {
...
    "/admin/audit-events": {
      "get": {
        "operationId": "listAuthorityAuditEvents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "events": {
                      "items": {
                        "$ref": "#/components/schemas/AuditEvent"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "events"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List bounded authority audit events",
        "tags": [
          "Administration"
        ],
        "x-codewithmee-contract": "listAuthorityAuditEvents",
        "parameters": [
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "pattern": "^(?:[1-9]|[1-9][0-9]|100)$",
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "listAuthorityUsers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "users": {
                      "items": {
                        "$ref": "#/components/schemas/AuthorityUser"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "users"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List bounded platform authority users",
        "tags": [
          "Administration"
        ],
        "x-codewithmee-contract": "listAuthorityUsers",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "pattern": "^(?:[1-9]|[1-9][0-9]|100)$",
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/users/{userId}/platform-role": {
      "patch": {
        "operationId": "changePlatformRole",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "auditEvent": {
                      "$ref": "#/components/schemas/AuditEvent"
                    },
                    "revokedSessionCount": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "user": {
                      "$ref": "#/components/schemas/AuthorityUser"
                    }
                  },
                  "required": [
                    "auditEvent",
                    "revokedSessionCount",
                    "user"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change a platform role with an audited revision check",
        "tags": [
          "Administration"
        ],
        "x-codewithmee-contract": "changePlatformRole",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{15,127}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "platformRole": {
                    "enum": [
                      "learner",
                      "moderator",
                      "superadmin",
                      "support"
                    ],
                    "type": "string"
                  },
                  "reason": {
                    "maxLength": 500,
                    "minLength": 12,
                    "type": "string"
                  },
                  "revision": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "platformRole",
                  "reason",
                  "revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/admin/users/{userId}/status": {
      "patch": {
        "operationId": "changeAccountStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "auditEvent": {
                      "$ref": "#/components/schemas/AuditEvent"
                    },
                    "revokedSessionCount": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "user": {
                      "$ref": "#/components/schemas/AuthorityUser"
                    }
                  },
                  "required": [
                    "auditEvent",
                    "revokedSessionCount",
                    "user"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change account status with an audited revision check",
        "tags": [
          "Administration"
        ],
        "x-codewithmee-contract": "changeAccountStatus",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{15,127}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "maxLength": 500,
                    "minLength": 12,
                    "type": "string"
                  },
                  "revision": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "status": {
                    "enum": [
                      "active",
                      "banned",
                      "deletion_pending",
                      "suspended"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "reason",
                  "revision",
                  "status"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/organizations/{organizationId}/ownership-transfer": {
      "post": {
        "operationId": "transferOrganizationOwnership",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "actorMembership": {
                      "$ref": "#/components/schemas/OwnershipMembership"
                    },
                    "auditEvent": {
                      "$ref": "#/components/schemas/AuditEvent"
                    },
                    "organization": {
                      "$ref": "#/components/schemas/OwnershipOrganization"
                    },
                    "targetMembership": {
                      "$ref": "#/components/schemas/OwnershipMembership"
                    }
                  },
                  "required": [
                    "actorMembership",
                    "auditEvent",
                    "organization",
                    "targetMembership"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Atomically transfer organization ownership",
        "tags": [
          "Administration",
          "Organizations"
        ],
        "x-codewithmee-contract": "transferOrganizationOwnership",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{15,127}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "maxLength": 500,
                    "minLength": 12,
                    "type": "string"
                  },
                  "revision": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "targetUserId": {
                    "maxLength": 100,
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
                    "type": "string"
                  }
                },
                "required": [
                  "reason",
                  "revision",
                  "targetUserId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}